// src/services/generated/models/predictionStatItem.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export interface PredictionStatItem {
  /** Total number of predictions */
  total_predictions: number;
  /** Number of malicious URLs detected */
  malicious_count: number;
  /** Number of safe URLs */
  safe_count: number;
  /** Number of pending predictions */
  pending_count: number;
  /** Overall accuracy rate */
  accuracy_rate: number;
  /** Average response time in ms */
  avg_response_time: number;
}