// src/services/generated/models/reportStatItem.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export interface ReportStatItem {
  /** ลิ้งค์ทีถูกรายงานโดยคุณ */
  reported_by_amount: number;
  /** ลิ้งค์ปลอดภัย */
  safe_urls: number;
  /** ลิ้งค์อันตราย */
  malicious_urls: number;
  /** รายงานผิดพลาด */
  error_urls: number;
  /** ลิ้งค์รอตรวจสอบ */
  pending_urls: number;
}